requirements.txt
streamlit>=1.31.0
azure-cognitiveservices-speech>=1.34.0
transformers>=4.36.0
torch>=2.0.0
python-dotenv>=1.0.0
streamlit-geolocation>=0.0.2
folium>=0.14.0
streamlit-folium>=0.15.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
pydub>=0.25.1
psutil>=5.9.0